Version 1.6
Added Content to aboutme.html, responsiveness to project.html and aboutme.html. Need to fix responsiveness for index.html, may have to do with .profile-section class. Updated resume to contain goobriel.github.io. Created .card class to give portfolio a bit more going on so it is not dull. .card sits behind texts, and images, and acts like a box filled with a light grey as opposed to background color.